==> 1.4-data_sci.ipynb <==
import pandas as pd
import json
import numpy as np
from collections import defaultdict
import matplotlib.pyplot as plt
import plotly.express as px
import seaborn as sns

df = pd.read_csv('data/processed.tsv', sep='\t')

df2 = pd.read_csv('data/ds-processed.tsv', sep='\t')

df2[(df2['degree'] == 'bachelor') & (df2['yoe_total'] < 2)]

df = df[:5].drop(['country', 'state'], axis=1)

df

df2 = df2[:5]

df.sample(frac=1)

df3 = pd.concat([df, df2], axis=0)

df3.reset_index(inplace=True, drop=True)
df3.sample(frac=1).reset_index(drop=True)

df2 = df2[df2['yoe_total'] <= 10]
df2 = df2[df2['degree'] != 'phd']
