==> 1.5-gay.ipynb <==
import pandas as pd
import json
import numpy as np
from collections import defaultdict
import matplotlib.pyplot as plt
import plotly.express as px
import seaborn as sns

df = pd.read_csv('data/processed.tsv', sep='\t')

df = df[df['location'] == 'San Francisco, CA']

pt = df.pivot_table(index='company', columns='gender', aggfunc='count', values='location')#.sort_values('Non-binary', ascending=False)[:10]
pt['total'] = pt.sum(axis=1)
pt['gay_pct'] = pt['Non-binary'] / pt['total']
pt['pct_techbros'] = pt['Male'] / pt['total']

pt[pt['Non-binary'] >= 3].sort_values('gay_pct', ascending=False)[:20]
pt[pt['total'] >= 25].sort_values('pct_techbros')[:20]

pt.loc[['Intel', 'Amazon']]

df[df['company'] == '7-Eleven']

grp.loc['Non-binary'].sort_values('location')
