==> 3.1-gensim_test.ipynb <==
from gensim.models.fasttext import FastText
import pandas as pd
import numpy as np

df = pd.read_csv('data/main-processed.tsv', sep='\t')

df['title'] = df['title'].str.lower().str.split()

df['log_tc'] = df['tc'].apply(lambda x: np.log(x))

from sklearn.pipeline import Pipeline
from sklearn.linear_model import LinearRegression
from sklearn.model_selection import train_test_split
from sklearn.compose import ColumnTransformer
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import OneHotEncoder

import category_encoders as ce

X = df.drop(['tc', 'base', 'stock', 'bonus', 'log_tc'], axis=1)
y = df['log_tc']

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=0)

model = FastText(vector_size=10)

corpus = X_train['title'].dropna()

corpus

model.build_vocab(corpus.values)

model.train(
    corpus_iterable=corpus, epochs=model.epochs,
    total_examples=model.corpus_count, total_words=model.corpus_total_words,
)

print(model)

wv = model.wv
wv.most_similar('junior')

# import gensim.downloader as api
# wv = api.load('word2vec-google-news-300')
